move appconfig into components folder and attempt to add paths
