+++ apps/sap-commerce-cloud/frontend/src/components/AppConfig/AppConfig.styles.ts
