+++ apps/sap-commerce-cloud/frontend/src/components/AppConfig/FieldSelector.styles.ts
