+++ apps/sap-commerce-cloud/frontend/src/components/AppConfig/fields.ts
