+++ apps/sap-commerce-cloud/frontend/src/components/AppConfig/parameters.ts
