+++ apps/sap-commerce-cloud/frontend/tsconfig.json
@@ -15,7 +15,15 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
-    "types": ["vite/client"]
+    "types": ["vite/client"], 
+    "paths": {
+      "@api/*": ["./src/api/*"],
+      "@components/*": ["./src/components/*"],
+      "@helpers/*": ["./src/helpers/*"],
+      "@hooks/*": ["./src/hooks/*"],
+      "@interfaces/*": ["./src/interfaces"],
+      "@__mocks__/*": ["./src/__mocks__/*"],
+    }
   },
   "include": ["src"]
 }
